// src/schemas/coupon/schema.ts
import { z } from "zod";
import { extendZodWithOpenApi } from "@asteasolutions/zod-to-openapi";

extendZodWithOpenApi(z);

export const AffiliateCouponSchema = z.object({
    id: z.string().uuid().openapi({
        description: "The unique identifier for the affiliate coupon.",
        example: "75434e84-255b-4314-a278-820df5e76813",
    }),
    external_id: z.string().openapi({
        description: "The external ID associated with the affiliate coupon, often used as a reference ID.",
        example: "promo_1a2b3c",
    }),
    token: z.string().openapi({
        description: "The unique token or code for the affiliate coupon.",
        example: "CODE",
    }),
    leads: z.number().int().openapi({
        description: "The number of leads generated by this coupon.",
        example: 0,
    }),
    conversions: z.number().int().openapi({
        description: "The number of conversions generated by this coupon.",
        example: 0,
    }),
    affiliate_id: z.string().uuid().openapi({
        description: "The ID of the affiliate associated with this coupon.",
        example: "95d48f70-f1d4-42d9-b929-21996b6d9eb4",
    }),
}).openapi("AffiliateCoupon");